Revert changes from main commits that were merged into blender-v4.1-release
The last good commit was 4bf6a2e564ad2.

+++ source/blender/draw/engines/eevee_next/shaders/eevee_vertex_copy_comp.glsl
@@ -5,7 +5,7 @@
 void main()
 {
   uint vertex_id = gl_GlobalInvocationID.x;
-  if (vertex_id >= uint(vertex_count)) {
+  if (vertex_id >= vertex_count) {
     return;
   }
   out_buf[start_offset + vertex_id] = vec4(in_buf[vertex_id * vertex_stride + 0],
